Reserve a straight path to end
Since the board is randomly spawned in, I have implemented code to clear the middle of the board, and the two z axis at the start and end which lead to the middle.

This could be modified in future to make sure the check only happens for walls and not guards.

Current potential issue with ReservedPointCheck() x and z are checked against hardcoded values.